+++ haskell-src/Concordium/Types/ProtocolVersion.hs
@@ -51,6 +51,7 @@ module Concordium.Types.ProtocolVersion (
     SomeProtocolVersion (..),
     promoteProtocolVersion,
     demoteProtocolVersion,
+    protocolVersionToWord64,
     MonadProtocolVersion (..),
 
     -- * Chain parameters version
@@ -158,6 +159,10 @@ module Concordium.Types.ProtocolVersion (
     --
     --    * 'TOV2' is used in P7 and onwards. The hash is computed similarly to 'TOV1',
     --      except the merkle trees are hashed to include the size.
+    --
+    --    * 'TOV3' is used in P10 and onwards. The hash is computed as before,
+    --      but the `TransactionSummary` contains a new `SponsorDetails` field
+    --      that is included in the hashing.
     TransactionOutcomesVersion (..),
     -- | Singleton type corresponding to 'TransactionOutcomesVersion'.
     STransactionOutcomesVersion (..),
@@ -169,6 +174,11 @@ module Concordium.Types.ProtocolVersion (
     TransactionOutcomesVersionFor,
     -- | The transaction outcomes version for a given protocol version (on singletons).
     sTransactionOutcomesVersionFor,
+    -- | Whether the `TransactionSummary` for the given transaction outcome
+    --  version contains the sponsor details field. Present since P10.
+    hasSponsorDetails,
+    sHasSponsorDetails,
+    HasSponsorDetails,
 
     -- * Delegation support
 
@@ -225,6 +235,13 @@ module Concordium.Types.ProtocolVersion (
     -- | Determine whether a specific protocol version supports protocol level tokens.
     PVSupportsPLT,
 
+    -- * Sponsored transactions support
+
+    -- | Determine whether sponsored transactions are supported.
+    SupportsSponsoredTransactions,
+    supportsSponsoredTransactions,
+    sSupportsSponsoredTransactions,
+
     -- * Block hash version
 
     -- | The version of the block hashing structure.
@@ -270,6 +287,11 @@ module Concordium.Types.ProtocolVersion (
     P7Sym0,
     P8Sym0,
     P9Sym0,
+    P10Sym0,
+    TOV0Sym0,
+    TOV1Sym0,
+    TOV2Sym0,
+    TOV3Sym0,
 ) where
 
 import Control.Monad.Except (ExceptT)
@@ -300,6 +322,7 @@ $( singletons
             | P7
             | P8
             | P9
+            | P10
             deriving (Eq, Ord)
 
         data ChainParametersVersion
@@ -319,6 +342,7 @@ $( singletons
         chainParametersVersionFor P7 = ChainParametersV2
         chainParametersVersionFor P8 = ChainParametersV3
         chainParametersVersionFor P9 = ChainParametersV3
+        chainParametersVersionFor P10 = ChainParametersV3
 
         -- \* Account versions
 
@@ -350,6 +374,7 @@ $( singletons
         accountVersionFor P7 = AccountV3
         accountVersionFor P8 = AccountV4
         accountVersionFor P9 = AccountV5
+        accountVersionFor P10 = AccountV5
 
         -- \|Authorizations version.
         data AuthorizationsVersion
@@ -368,6 +393,7 @@ $( singletons
         authorizationsVersionFor P7 = AuthorizationsVersion1
         authorizationsVersionFor P8 = AuthorizationsVersion1
         authorizationsVersionFor P9 = AuthorizationsVersion2
+        authorizationsVersionFor P10 = AuthorizationsVersion2
 
         -- \|Transaction outcomes versions.
         -- The difference between the two versions are only related
@@ -378,10 +404,14 @@ $( singletons
         --  exact reject reasons for failed transactions are omitted from the hash.
         --  * 'TOV2' is used in P7 and onwards. The hash is computed similarly to 'TOV1',
         --  except the merkle trees are hashed to include the size.
+        --  * 'TOV3' is used in P10 and onwards. The hash is computed as before,
+        --    but the `TransactionSummary` contains a new `SponsorDetails` field
+        --    that is included in the hashing.
         data TransactionOutcomesVersion
             = TOV0
             | TOV1
             | TOV2
+            | TOV3
 
         -- \|Projection of 'ProtocolVersion' to 'TransactionOutcomesVersion'.
         transactionOutcomesVersionFor :: ProtocolVersion -> TransactionOutcomesVersion
@@ -394,6 +424,15 @@ $( singletons
         transactionOutcomesVersionFor P7 = TOV2
         transactionOutcomesVersionFor P8 = TOV2
         transactionOutcomesVersionFor P9 = TOV2
+        transactionOutcomesVersionFor P10 = TOV3
+
+        -- \| Whether the `TransactionSummary` for the given transaction outcome
+        --  version contains the sponsor details field. Present since P10.
+        hasSponsorDetails :: TransactionOutcomesVersion -> Bool
+        hasSponsorDetails TOV0 = False
+        hasSponsorDetails TOV1 = False
+        hasSponsorDetails TOV2 = False
+        hasSponsorDetails TOV3 = True
 
         supportsDelegation :: AccountVersion -> Bool
         supportsDelegation AccountV0 = False
@@ -427,6 +466,18 @@ $( singletons
         supportsPLT AccountV4 = False
         supportsPLT AccountV5 = True
 
+        supportsSponsoredTransactions :: ProtocolVersion -> Bool
+        supportsSponsoredTransactions P1 = False
+        supportsSponsoredTransactions P2 = False
+        supportsSponsoredTransactions P3 = False
+        supportsSponsoredTransactions P4 = False
+        supportsSponsoredTransactions P5 = False
+        supportsSponsoredTransactions P6 = False
+        supportsSponsoredTransactions P7 = False
+        supportsSponsoredTransactions P8 = False
+        supportsSponsoredTransactions P9 = False
+        supportsSponsoredTransactions P10 = True
+
         -- \| A type representing the different hashing structures used for the block hash depending on
         -- the protocol version.
         data BlockHashVersion
@@ -446,6 +497,7 @@ $( singletons
         blockHashVersionFor P7 = BlockHashVersion1
         blockHashVersionFor P8 = BlockHashVersion1
         blockHashVersionFor P9 = BlockHashVersion1
+        blockHashVersionFor P10 = BlockHashVersion1
 
         -- \| Whether the block state hash is tracked as part of the block metadata.
         blockStateHashInMetadata :: BlockHashVersion -> Bool
@@ -472,6 +524,7 @@ protocolVersionToWord64 P6 = 6
 protocolVersionToWord64 P7 = 7
 protocolVersionToWord64 P8 = 8
 protocolVersionToWord64 P9 = 9
+protocolVersionToWord64 P10 = 10
 
 -- | Parse a 'Word64' as a 'ProtocolVersion'.
 protocolVersionFromWord64 :: (MonadFail m) => Word64 -> m ProtocolVersion
@@ -484,6 +537,7 @@ protocolVersionFromWord64 6 = return P6
 protocolVersionFromWord64 7 = return P7
 protocolVersionFromWord64 8 = return P8
 protocolVersionFromWord64 9 = return P9
+protocolVersionFromWord64 10 = return P10
 protocolVersionFromWord64 v = fail $ "Unknown protocol version: " ++ show v
 
 -- | Convert a @ChainParametersVersion@ to the corresponding 'Word64'.
@@ -521,6 +575,7 @@ promoteProtocolVersion P6 = SomeProtocolVersion SP6
 promoteProtocolVersion P7 = SomeProtocolVersion SP7
 promoteProtocolVersion P8 = SomeProtocolVersion SP8
 promoteProtocolVersion P9 = SomeProtocolVersion SP9
+promoteProtocolVersion P10 = SomeProtocolVersion SP10
 
 -- | Demote an 'SProtocolVersion' to a 'ProtocolVersion'.
 demoteProtocolVersion :: SProtocolVersion pv -> ProtocolVersion
@@ -718,6 +773,7 @@ supportsMemo SP6 = True
 supportsMemo SP7 = True
 supportsMemo SP8 = True
 supportsMemo SP9 = True
+supportsMemo SP10 = True
 
 -- | Whether the protocol version supports account aliases.
 --  (Account aliases are supported in 'P3' onwards.)
@@ -731,6 +787,7 @@ supportsAccountAliases SP6 = True
 supportsAccountAliases SP7 = True
 supportsAccountAliases SP8 = True
 supportsAccountAliases SP9 = True
+supportsAccountAliases SP10 = True
 
 -- | Whether the protocol version supports V1 smart contracts.
 --  (V1 contracts are supported in 'P4' onwards.)
@@ -744,6 +801,7 @@ supportsV1Contracts SP6 = True
 supportsV1Contracts SP7 = True
 supportsV1Contracts SP8 = True
 supportsV1Contracts SP9 = True
+supportsV1Contracts SP10 = True
 
 -- | Whether the protocol version supports delegation.
 --  (Delegation is supported in 'P4' onwards.)
@@ -757,6 +815,7 @@ supportsDelegationPV SP6 = True
 supportsDelegationPV SP7 = True
 supportsDelegationPV SP8 = True
 supportsDelegationPV SP9 = True
+supportsDelegationPV SP10 = True
 
 -- | Whether the protocol version supports upgradable smart contracts.
 --  (Supported in 'P5' and onwards)
@@ -771,6 +830,7 @@ supportsUpgradableContracts spv = case spv of
     SP7 -> True
     SP8 -> True
     SP9 -> True
+    SP10 -> True
 
 -- | Whether the protocol version supports chain queries in smart contracts.
 --  (Supported in 'P5' and onwards)
@@ -785,6 +845,7 @@ supportsChainQueryContracts spv = case spv of
     SP7 -> True
     SP8 -> True
     SP9 -> True
+    SP10 -> True
 
 -- | Whether the protocol version supports sign extension instructions for V1
 --  contracts. (Supported in 'P6' and onwards)
@@ -799,6 +860,7 @@ supportsSignExtensionInstructions spv = case spv of
     SP7 -> True
     SP8 -> True
     SP9 -> True
+    SP10 -> True
 
 -- | Whether the protocol version allows globals in data and element sections of
 --  Wasm modules for V1 contracts. (Supported before 'P6')
@@ -813,6 +875,7 @@ supportsGlobalsInInitSections spv = case spv of
     SP7 -> False
     SP8 -> False
     SP9 -> False
+    SP10 -> False
 
 -- | Whether the protocol version specifies that custom section should not be
 --  counted towards module size when executing V1 contracts.
@@ -833,6 +896,7 @@ supportsAccountSignatureChecks spv = case spv of
     SP7 -> True
     SP8 -> True
     SP9 -> True
+    SP10 -> True
 
 -- | Whether the protocol version supports querying a smart contract's module reference and name
 --  from smart contracts.
@@ -848,6 +912,7 @@ supportsContractInspectionQueries = \case
     SP7 -> True
     SP8 -> True
     SP9 -> True
+    SP10 -> True
 
 -- | Whether the protocol version supports encrypting balances and sending encrypted transfers.
 --  (Disabled in 'P7' and onwards.)
@@ -862,3 +927,4 @@ supportsEncryptedTransfers = \case
     SP7 -> False
     SP8 -> False
     SP9 -> False
+    SP10 -> False