+++ haskell-tests/Types/TransactionSummarySpec.hs
@@ -54,6 +54,7 @@ tests = describe "Transaction summaries" $ do
     versionedTests SP7
     versionedTests SP8
     versionedTests SP9
+    versionedTests SP10
   where
     versionedTests spv = describe (show $ demoteProtocolVersion spv) $ do
         specify "Event: serialize then deserialize is identity" $ withMaxSuccess 10000 $ testEventSerializationIdentity spv